starting events

+++ main.py
@@ -1,4 +1,5 @@
 import datetime
+import random
 import io
 import time
 from io import BytesIO
@@ -7,6 +8,7 @@ from flask_mail import Mail, Message
 from flask_sqlalchemy import SQLAlchemy
 from flask_login import UserMixin, login_user, LoginManager, login_required, logout_user, current_user
 import json
+import sqlalchemy
 from sqlalchemy.sql import func
 from werkzeug.utils import secure_filename
 import Search_Engine
@@ -16,6 +18,8 @@ from flask_wtf.csrf import CSRFProtect
 import sys
 from profanity_check import predict
 import smtplib
+from flask_socketio import join_room,leave_room,send,SocketIO
+import functions
 
 sender_email = "anonymousrobot974@gmail.com"
 
@@ -100,12 +104,11 @@ class Teacher(db.Model, UserMixin):
     Username = db.Column(db.String(100), nullable=False, unique=True)
     Email = db.Column(db.String(100), nullable=False, unique=True)
     Mobile = db.Column(db.String(100), nullable=False, unique=True)
-    Class = db.Column(db.String(100), nullable=False)
-    Salary = db.Column(db.Integer, nullable=False)
     Password = db.Column(db.String(100), nullable=False)
-    Class1 = db.Column(db.String(100), nullable=False)
-    Class2 = db.Column(db.String(100), nullable=False)
-    Class3 = db.Column(db.String(100), nullable=False)
+    Class = db.Column(db.String(100))
+    Class1 = db.Column(db.String(100))
+    Class2 = db.Column(db.String(100))
+    Class3 = db.Column(db.String(100))
 
     def get_id(self):
         return self.TID
@@ -147,6 +150,17 @@ class Notes(db.Model,UserMixin):
         return self.NID
 
 
+class Chat(db.Model,UserMixin):
+    chat_id = db.Column(db.Integer, primary_key=True)
+    code = db.Column(db.String(100), nullable=False)
+    created = db.Column(db.String(200), nullable=False)
+    
+    def get_id(self):
+        return self.chat_id
+    
+socketio = SocketIO(app)
+
+
 @app.route('/', methods=['GET', 'POST'])
 def login_page():
     form = forms.LoginForm()
@@ -170,6 +184,50 @@ def login_page():
             error = "No such user"
 
     return render_template('login.html', error=error, form=form)
+
+
+
+@app.route('/Signup', methods=['GET', 'POST'])
+def Signup_page():
+    form = forms.SignUpForm()
+    error = ""
+    if form.validate_on_submit():
+        name = form.Name.data
+        email = form.Email.data
+        member = form.member.data
+        mobile = form.Mobile.data
+        password = bcrypt.generate_password_hash(form.Password.data)
+
+        # try:
+        if member == 'teacher':
+                if email[-19:] == "christuniversity.in" and email[email.find("@")+1:-20] in ['science']:
+                    new_member = Teacher(Name=name, Mobile=mobile,Username=mobile,Class="",Class1="",Class2="",Class3="", Password=password, Email=email)
+                    db.session.add(new_member)
+                    db.session.commit()
+                    return redirect(url_for('login_page'))
+                else:
+                    error = "Not a valid email"
+        elif member == 'student':
+            # print(email[email.find("@")+1:-20])
+            if email[-19:] == "christuniversity.in" and email[email.find("@")+1:-20] in ['bca','ems','cms','bda']:
+                new_member = Student(Name=name, Mobile=mobile,Username=mobile,Class="1"+str(email[email.find("@")+1:-20]), Password=password, Email=email)
+                db.session.add(new_member)
+                db.session.commit()
+                return redirect(url_for('login_page'))
+            else:
+                error = "Not a valid email"
+
+        # except sqlalchemy.exc.IntegrityError:
+        #     error = "user already there"
+        # except:
+        #     error = "unknown error occured"
+
+
+
+
+    return render_template('signup.html', error=error, form=form)
+
+
 
 
 @app.route('/admin-page', methods=['GET', 'POST'])
@@ -213,26 +271,50 @@ def student_login():
         return redirect(url_for('login_page'))
     form = forms.PersonalDetails()
 
-    class_name = user = getattr(sys.modules[__name__], session['user'].capitalize())
+    class_name = getattr(sys.modules[__name__], session['user'].capitalize())
     personal_class_name = getattr(sys.modules[__name__], f"{session['user'].capitalize()}Personal")
     details = class_name.query.filter_by(Username=session['Username']).first()
+
     all_user_details = db.session.query(class_name, personal_class_name).filter_by(Username=session['Username']) \
         .outerjoin(personal_class_name, class_name.Username == personal_class_name.Username).first()
+
     if all_user_details[1] == None:
         if form.validate_on_submit():
-            new = personal_class_name(Username=details.Username, Marks_10=form.Marks_10.data,
-                                      Marks_12=form.Marks_12.data
-                                      , City=form.City.data, State=form.State.data, Country=form.Country.data,
-                                      Local_Guardian=form.Local_Guardian.data, Father_Name=form.Father_Name.data,
-                                      Mother_Name=form.Mother_Name.data)
+            new = personal_class_name(Username=details.Username, Marks_10=form.Marks_10.data,Marks_12=form.Marks_12.data, City=form.City.data, State=form.State.data, Country=form.Country.data,Local_Guardian=form.Local_Guardian.data, Father_Name=form.Father_Name.data,Mother_Name=form.Mother_Name.data)
+            
+
             db.session.add(new)
             db.session.commit()
+
+
             return redirect(url_for('student_login'))
         return render_template('personal_details.html', form=form, member=session['user'])
 
-    announcement =  db.session.query(class_name,Announcement).filter_by(Class=details.Class).join(class_name,Announcement.Class==class_name.Class).all()
+    announcement =  db.session.query(class_name,Announcement).filter_by(Class=details.Class).join(class_name,Announcement.Class==class_name.Class).order_by(Announcement.Date_time.desc()).all()
 
-    return render_template('student_index.html', name=name, user=all_user_details,announcements=announcement)
+    for i in announcement:
+        if (datetime.datetime.today() -  i[1].Date_time).days > 1 :
+            db.session.delete(i[1])
+            db.session.commit()
+
+    announcement_list = []
+
+    for i in announcement:
+        announcement_list.append(f"{i[1].TEXT} ------- {i[1].TName} ({i[1].Date_time})")
+
+    announcement_list = [*set(announcement_list)]
+
+    chatform = forms.Chat()
+    if chatform.validate_on_submit():
+        code = chatform.name.data
+        data = Chat.query.filter_by(code=code).first()
+        if data != None:
+            print("Chat exist")
+        return redirect(url_for('chat',code=code,type="join"))
+
+
+    return render_template('student_index.html', name=name, user=all_user_details,announcements=announcement_list,chatform=chatform)
+
 
 @app.route('/download-notes')
 @login_required
@@ -248,15 +330,12 @@ def view_notes():
     student_query = Student.query.filter_by(Username=session['Username']).first()
     student_class = student_query.Class
     count_notes_class = Notes.query.filter_by(Class=student_class).count()
-    print(count_notes_class)
+    # print(count_notes_class)
     all_notes = db.session.query(Student, Notes).filter_by(Class=student_class)\
         .join(Notes, Student.Class == Notes.Class).all()
     first_notes = all_notes[:count_notes_class]
 
     return render_template('download_notes.html',notes=first_notes)
-
-
-
 
 
 @app.route('/upload-notes',methods=['GET','POST'])
@@ -276,20 +355,20 @@ def upload_notes():
     if form.validate_on_submit():
         data = form.file.data
         selected_class = request.form.get('classes')
-        upload = Notes(data=data.read(),Message=form.description.data,Class=selected_class,TName="Messi",type=data.mimetype,filename=data.filename)
+        upload = Notes(data=data.read(),Message=form.description.data,Class=selected_class,TName=user.Name,type=data.mimetype,filename=data.filename)
         db.session.add(upload)
-        # db.session.commit()
+        db.session.commit()
         Students = Student.query.filter_by(Class=selected_class).all()
         msg = Message(
             'Material Uploaded !!!!',
             sender=sender_email,
             recipients=[i.Email for i in Students]
         )
-        print(msg.recipients)
-        msg.body = f'Dear Students material has been uploaded in your portal'
-        print(data.mimetype)
-        print(data.filename)
-        msg.attach(data.filename, data.mimetype, data.read())
+        # print(msg.recipients)
+        msg.body = f'Dear Students material has been uploaded in your portal by {user.Name} - {data.filename}'
+        # print(data.mimetype)
+        # print(data.filename)
+        # msg.attach(data.filename, data.mimetype, data.read())
         mail.send(msg)
         return redirect(url_for('teacher_login'))
     return render_template('notes.html',form=form,user=user)
@@ -297,7 +376,7 @@ def upload_notes():
 @app.route('/download')
 def download():
     notes_id = request.args.get('id')
-    print(notes_id)
+    # print(notes_id)
     upload = Notes.query.filter_by(NID=notes_id).first()
     return send_file(io.BytesIO(upload.data),as_attachment=True,download_name=upload.filename,mimetype=upload.type)
 
@@ -313,18 +392,35 @@ def teacher_login():
             return redirect(url_for('login_page'))
     else:
         return redirect(url_for('login_page'))
-    class_name = user = getattr(sys.modules[__name__], session['user'].capitalize())
+    class_name = getattr(sys.modules[__name__], session['user'].capitalize())
     personal_class_name = getattr(sys.modules[__name__], f"{session['user'].capitalize()}Personal")
+    details = class_name.query.filter_by(Username=session['Username']).first()
     all_user_details = db.session.query(class_name, personal_class_name).filter_by(Username=session['Username']) \
         .outerjoin(personal_class_name, class_name.Username == personal_class_name.Username).first()
+    p_form = forms.TeacherPersonalDetails()
+
+    # print(all_user_details[1])
+
+    if all_user_details[1] == None:
+        if p_form.validate_on_submit():
+            new = personal_class_name(Username=details.Username,  City=p_form.City.data, State=p_form.State.data, Country=p_form.Country.data,Qualification=p_form.qualification.data)
+            
+
+            db.session.add(new)
+            db.session.commit()
+
+
+            return redirect(url_for('teacher_login'))
+        return render_template('personal_details.html', form=p_form, member=session['user'])
+
     form = forms.announcement()
     if form.validate_on_submit():
         text = form.announcement.data
         cur_time = func.now()
-        print(cur_time)
+        # print(cur_time)
         Class = request.form.get('classes')
         tname = session['user-name']
-        print(tname)
+        # print(tname)
         new = Announcement(TEXT=text,Class=Class,Date_time=cur_time,TName=tname)
         db.session.add(new)
         db.session.commit()
@@ -347,8 +443,13 @@ def add_class():
 
     form = forms.add_class()
     teacher = Teacher.query.filter_by(Username=session['Username']).first()
+    # print(teacher.Class+" yo")
     if form.validate_on_submit():
-        if teacher.Class1 == "":
+        if teacher.Class == "":
+            teacher.Class = form.SEM.data + form.Course.data
+            db.session.commit()
+
+        elif teacher.Class1 == "":
             teacher.Class1 = form.SEM.data + form.Course.data
             if teacher.Class1 == teacher.Class:
                 error = "teacher is already teaching this class"
@@ -411,13 +512,11 @@ def update_member():
     form = forms.UpdateMember()
     cur_member = getattr(sys.modules[__name__], session['to-update'].capitalize()).query.filter_by(
         Username=session['Username']).first()
-    if session['to-update'] == 'Teacher':
-        salary_value = cur_member.Salary
+
     sem_value = cur_member.Class[0]
     mobile_value = cur_member.Mobile
     email_value = cur_member.Email
     if form.validate_on_submit():
-        cur_member.Email = form.Email.data
         cur_member.Class = form.SEM.data + cur_member.Class[1:]
         cur_member.Mobile = form.Mobile.data
         if session['to-update'] == 'Teacher':
@@ -436,7 +535,8 @@ def delete_member():
     user = getattr(sys.modules[__name__], member).query.filter_by(Username=username).first()
     personal_user = getattr(sys.modules[__name__], f"{member}Personal").query.filter_by(Username=username).first()
     db.session.delete(user)
-    db.session.delete(personal_user)
+    if personal_user != None:
+        db.session.delete(personal_user)
     db.session.commit()
     return redirect(url_for('admin_login'))
 
@@ -476,7 +576,7 @@ def create_member():
         db.session.add(new_member)
         db.session.commit()
         return redirect(url_for('admin_login'))
-    print('not validated')
+    # print('not validated')
     return render_template('make-member.html', form=form, member=member)
 
 
@@ -500,6 +600,9 @@ def member_details():
     session['to-update'] = session['member']
     return render_template('member_details.html', user=all_user_details, member=session['member'])
 
+@app.route("/Events/send")
+def upload_events():
+    return render_template
 
 @app.route('/logout')
 @login_required
@@ -513,14 +616,16 @@ def logout():
 if __name__ == '__main__':
     with app.app_context():
         db.create_all()
-    app.run(debug=True)
+    socketio.run(app=app,debug=True)
+
+
 
 # @app.route('/make-admin', methods=['GET', 'POST'])
 # def make_admin():
 #     form = forms.MakeAdmin()
 #     if form.validate_on_submit():
 #         username = form.Username.data
-#         password = bcrypt.generate_password_hash(form.Password.data).decode('utf-8')
+#         password = bcrypt.generate_password_hash(form.Password.data decode('utf-8')
 #         name = form.Name.data
 #         new = Admin(Name=name, Username=username, Password=password)
 #         db.session.add(new)
